+++ test/test_modal_manager.ml
@@ -37,6 +37,18 @@ module Modal_page = struct
     if key = "inc" then Miaou_core.Navigation.update (fun st -> st + 1) ps
     else ps
 
+  let on_key ps key ~size =
+    let key_str = Miaou_core.Keys.to_string key in
+    let ps' = handle_key ps key_str ~size in
+    (ps', Miaou_interfaces.Key_event.Bubble)
+
+  let on_modal_key ps key ~size =
+    let key_str = Miaou_core.Keys.to_string key in
+    let ps' = handle_modal_key ps key_str ~size in
+    (ps', Miaou_interfaces.Key_event.Bubble)
+
+  let key_hints _ = []
+
   let has_modal _ = false
 end