+++ test/test_registry_validation.ml
@@ -36,6 +36,18 @@ module Good_page : Miaou_core.Tui_page.PAGE_SIG = struct
 
   let handle_key ps _ ~size:_ = ps
 
+  let on_key ps key ~size =
+    let key_str = Miaou_core.Keys.to_string key in
+    let ps' = handle_key ps key_str ~size in
+    (ps', Miaou_interfaces.Key_event.Bubble)
+
+  let on_modal_key ps key ~size =
+    let key_str = Miaou_core.Keys.to_string key in
+    let ps' = handle_modal_key ps key_str ~size in
+    (ps', Miaou_interfaces.Key_event.Bubble)
+
+  let key_hints _ = []
+
   let has_modal _ = false
 end
 
@@ -75,6 +87,18 @@ module Bad_page : Miaou_core.Tui_page.PAGE_SIG = struct
   let handle_modal_key ps _ ~size:_ = ps
 
   let handle_key ps _ ~size:_ = ps
+
+  let on_key ps key ~size =
+    let key_str = Miaou_core.Keys.to_string key in
+    let ps' = handle_key ps key_str ~size in
+    (ps', Miaou_interfaces.Key_event.Bubble)
+
+  let on_modal_key ps key ~size =
+    let key_str = Miaou_core.Keys.to_string key in
+    let ps' = handle_modal_key ps key_str ~size in
+    (ps', Miaou_interfaces.Key_event.Bubble)
+
+  let key_hints _ = []
 
   let has_modal _ = false
 end
@@ -120,6 +144,18 @@ module Conflicting_page : Miaou_core.Tui_page.PAGE_SIG = struct
   let handle_modal_key ps _ ~size:_ = ps
 
   let handle_key ps _ ~size:_ = ps
+
+  let on_key ps key ~size =
+    let key_str = Miaou_core.Keys.to_string key in
+    let ps' = handle_key ps key_str ~size in
+    (ps', Miaou_interfaces.Key_event.Bubble)
+
+  let on_modal_key ps key ~size =
+    let key_str = Miaou_core.Keys.to_string key in
+    let ps' = handle_modal_key ps key_str ~size in
+    (ps', Miaou_interfaces.Key_event.Bubble)
+
+  let key_hints _ = []
 
   let has_modal _ = false
 end